refactor planner helper extraction
